specify unstable features in compiler options
previously, i was getting insane errors like "`#![feature]` may not be
used on the nightly channel" because we filled in the
`unstable_features` field with
`rustc_session::config::Options::default()`, which does not have any
logic to detect a nightly toolchain. with this patch, it notices that it
is nightly, and doesn't error on nightly features.